+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Add
 + Add Frontend version on store
++ Add Accounts table on frontend
 
 ## 0.1.0
 + First version